feat: adiciona endpoint bruno

+++ diacono/src/main/java/com/diacono/diacono/dashboard/model/response/membro/DashboardGeneroMembroDTO.java
@@ -1,7 +1,7 @@
 package com.diacono.diacono.dashboard.model.response.membro;
 
 public record DashboardGeneroMembroDTO(
-        long masculino,
-        long feminino
+        double masculino,
+        double feminino
 ){
 }
